@@ -4,77 +4,76 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "build": "ts-node fuse.ts",
-    "prod": "cross-env NODE_ENV=production npm run build",
-    "start": "npm run -s build -- -d -h -t",
-    "dev": "npm run -s build -- -d",
-    "watch": "npm run -s build -- -w",
-    "test": "mocha",
-    "cover": "nyc mocha"
+    "dev": "vite",
+    "build": "vite build && rm -rf docs/assets && rsync -ahP dist/ docs/",
+    "preview": "vite preview",
+    "prod": "cross-env NODE_ENV=production npm run build && rm -rf docs/assets && rsync -ahP dist/ docs/",
+    "start": "npm run -s dev",
+    "watch": "npm run -s dev"
   },
   "devDependencies": {
-    "@types/chai": "^4.3.0",
+    "@types/chai": "^4.3.4",
     "@types/chai-enzyme": "^0.6.9",
     "@types/classnames": "^2.3.0",
-    "@types/enzyme": "^3.10.11",
-    "@types/glob": "^7.2.0",
-    "@types/history": "^4.7.9",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^17.0.8",
+    "@types/enzyme": "^3.10.12",
+    "@types/glob": "^8.0.0",
+    "@types/history": "^4.7.11",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.11.18",
     "@types/node-fetch": "^3.0.2",
-    "@types/prop-types": "^15.7.4",
+    "@types/prop-types": "^15.7.5",
     "@types/query-string": "^6.3.0",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.21",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
-    "@types/react-router-redux": "^5.0.20",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
     "@types/recompose": "^0.30.10",
     "@types/redux-immutable-state-invariant": "^2.1.2",
     "@types/redux-logger": "^3.0.9",
-    "@types/sinon": "^10.0.6",
-    "chai": "^4.3.4",
-    "chai-enzyme": "^0.8.0",
-    "cheerio": "^1.0.0-rc.10",
+    "@types/sinon": "^10.0.13",
+    "@vitejs/plugin-react": "^3.0.1",
+    "chai": "^4.3.7",
+    "cheerio": "^1.0.0-rc.12",
     "cross-env": "^7.0.3",
     "cue-parser-plus": "^0.0.5",
     "enzyme": "^3.11.0",
     "fuse-box": "^4.0.1-next.4",
-    "mocha": "^9.1.3",
+    "mocha": "^10.2.0",
     "nyc": "^15.1.0",
-    "sinon": "^12.0.1",
-    "ts-node": "^10.4.0",
-    "tslib": "^2.3.1",
-    "typescript": "^4.5.4",
-    "uglify-js": "^3.14.5"
+    "sinon": "^15.0.1",
+    "ts-node": "^10.9.1",
+    "tslib": "^2.4.1",
+    "typescript": "^4.9.4",
+    "uglify-js": "^3.17.4",
+    "vite": "^4.0.4"
   },
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
-    "@babel/runtime": "^7.16.7",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "classnames": "^2.3.1",
-    "fetch-jsonp": "^1.2.1",
-    "history": "^5.2.0",
-    "jss-preset-default": "^10.9.0",
-    "preact": "^10.6.4",
-    "preact-compat": "^3.19.0",
-    "preact-context": "^1.1.4",
-    "react": "^17.0.2",
+    "@babel/runtime": "^7.20.7",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/lab": "^5.0.0-alpha.114",
+    "@mui/material": "^5.11.3",
+    "@mui/styles": "^5.11.2",
+    "classnames": "^2.3.2",
+    "fetch-jsonp": "^1.2.3",
+    "history": "^5.3.0",
+    "jss-preset-default": "^10.9.2",
+    "react": "^18.2.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^17.0.2",
-    "react-jss": "^10.9.0",
-    "react-redux": "^7.2.6",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
-    "react-router-redux": "5.0.0-alpha.6",
-    "redux": "^4.1.2",
+    "react-dom": "^18.2.0",
+    "react-jss": "^10.9.2",
+    "react-redux": "^8.0.5",
+    "react-router": "^6.6.1",
+    "react-router-dom": "^6.6.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-immutable-state-invariant": "^2.1.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.4.1",
-    "reselect": "^4.1.5",
+    "redux-thunk": "^2.4.2",
+    "reselect": "^4.1.7",
     "smoothscroll-polyfill": "^0.4.4",
     "typescript-fsa": "^3.0.0",
     "typescript-fsa-reducers": "^1.2.2",
