@@ -1,36 +1,34 @@
-import * as React from 'react';
+import * as React from "react";
 
-type PreloadProps = {
-    preload: () => any;
-    wait?: boolean;
+interface PreloadProps extends React.PropsWithChildren {
+  preload: () => any;
+  wait?: boolean;
 }
 
 type PreloadState = {
-    renderOK: boolean;
-}
+  renderOK: boolean;
+};
 
 export class Preload extends React.Component<PreloadProps, PreloadState> {
-    state = { renderOK: false };
+  state = { renderOK: false };
 
-    async preload() {
-        await this.setState({ renderOK: !this.props.wait });
-        await this.props.preload();
-        await this.setState({ renderOK: true });
-    }
+  async preload() {
+    await this.setState({ renderOK: !this.props.wait });
+    await this.props.preload();
+    await this.setState({ renderOK: true });
+  }
 
-    renderPreloaded() {
-        return this.state.renderOK
-            ? this.props.children
-            : [];
-    }
+  renderPreloaded() {
+    return this.state.renderOK ? this.props.children : [];
+  }
 
-    componentDidMount() {
-        this.preload();
-    }
+  componentDidMount() {
+    this.preload();
+  }
 
-    render() {
-        return <div>{this.renderPreloaded()}</div>;
-    }
+  render() {
+    return <div>{this.renderPreloaded()}</div>;
+  }
 }
 
 export default Preload;
